// src/interfaces/IUser.ts
export interface IUser {
  firstName: string | null;
  lastName: string | null;
  displayName: string | null;
  email: string | null;
  gender: string | null;
  location: string | null;
  number: string | null;
  country: string | null;
  birthday: string | null;
  // password: string;
  // confirmPassword: string;
}